Initial commit
This is a initial commit of version 0.0.0. It should have been a commit
of version 0.1.0 which was much further and contained almost fully
implemented user login. However, it was a Friday, the 13th and my
attempt to create a new repository and push the program went so bad that
I ended up with the whole code deleted and with no backup. The only
backup I've managed to find is this one which is three months old.